Add files via upload
We will pick the closing prices of 5 stocks (Apple, Microsoft, JP Morgan, Deutsche Bank and Gold as a commodity). The idea behind mean-variance portfolio theory (MPT) is that we want to diversify our investments, which involves in either minimising risk for fixed expected return from our portfolio or maximise the returns from the portfolio for fixed risk.

Later, we are going to use Monte Carlo simulation to simulate stock prices given by the following stochastic differential equation in the Black-Scholes-Merton setup.